Created: Consent model, Consent service, Privacy policy model, Consent dao and Privacy policy dao. Updated Service and Userdao to remove foreign key references before actual deletion. Updated tests to affect new db structure. Created migrations for privacy policy and user consent. Created seed for privacy policies.

+++ src/dao/UserDao.ts
@@ -179,9 +179,15 @@ export default class UserDao implements IDao<IUserDatabaseObject> {
    * @memberof UserDao
    */
   public remove(id: number): Promise<boolean> {
-    return this.knex("users")
+    // First, delete consents
+    return this.knex("privacy_policy_consent_data")
       .delete()
-      .where({ id });
+      .where({ user_id: id })
+      .then((result: any) => {
+        return this.knex("users")
+          .delete()
+          .where({ id });
+      });
   }
 
   /**
